discord-bot/requirements.txt
discord.py>=2.0.0
aiohttp>=3.8.0
gunicorn==20.1.0
